refactor: add memo
