+++ react-game/src/index.css
@@ -109,6 +109,10 @@ body, html, .App {
     height: 100%;
     width: calc(300px + (700 - 300) * ((100vw - 320px) / (1400 - 320)));
 }
+.settings-buttons{
+    text-align: center;
+}
+
 .menu-content{
     line-height: 36px;
     flex-direction: column;
@@ -135,7 +139,7 @@ body, html, .App {
     width: 70%;
 }
 .label{
-    font-size: calc(28px + (32 - 28) * ((100vw - 320px) / (1400 - 320)))
+    font-size: calc(24px + (28 - 24) * ((100vw - 320px) / (1400 - 320)))
 }
 /*  INPUT =========================================================================================================== */
 .label-input{
@@ -157,6 +161,7 @@ body, html, .App {
     margin: 10px;
     display: flex;
     flex-direction: column;
+    display: block;
 }
 .hotkeys button{
     width: 150px;
@@ -306,6 +311,15 @@ footer b{
 
 @media (max-height: 750px) {
     .hotkeys{
-
+        display: none;
+    }
+    .settings-buttons{
+        width: 80%;
+        justify-content: space-between;
+        display: flex;
+    }
+    .settings-buttons .settings-button{
+        display: flex;
+        flex-direction: column;
     }
 }
